requirements.txt
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.1
psycopg2==2.8.5
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.8
vk-api==11.8.0
wrapt==1.11.2